ecommerce/requirements.txt
duckdb
pandas
pymysql
sqlalchemy
faker
python-dotenv
dbt-core
dbt-duckdb
apache-airflow-providers-mysql